Adding a pre-commit hook for the master branch to run all tests
Closes #39

This adds the pre-commit library to this repo so we can run a script
when someone tries to commit to the master branch -- this script runs
our automated tests and if they fail, won't allow the user to commit
their changes to the master branch. This only works when squash merging
locally, as opposed to on GitHub, but will help regardless.

Reviewers: @gabesmed @marimiyachi