QPPA-1409: Update benchmarks from HCQAR file
Closes #104

https://jira.cms.gov/browse/QPPA-1409

This PR adds in changes to the benchmarks from an HCQAR file (see
QPPA-1409). The only modifications made to the file was changing the
column names to match what scripts/benchmarks/parse-benchmarks-data.js
expects and to export the data as a CSV (since it came to us as an excel
file). I then placed that file into staging/benchmarks/csv/ and ran:

cat staging/benchmarks/csv/003_benchmarks_hcqar_updates.csv | node
scripts/benchmarks/parse-benchmarks-data.js 2015 2017
to generate the JSON. Then I ran npm run build:benchmarks, which updated
the 2017.json benchmarks file.

I also noticed a small mistake in the CONTRIBUTING.md guide, so I've
fixed that in this PR.

Lastly, bumping the package version number from 1.0.1 to 1.0.2.

Here is a screenshot of the two relevant lines from the original CSV to
compare against the changes to 2017.json generated:
screen shot 2018-01-03 at 1 00 22 pm
The changes in 2017.json look sensible in light of this data -- seems
like the parse-benchmarks-data.js script added a first decile of 100 for
ACEP21 and 0 for ACEP24 in addition to what was in the file (though it
didn't add a decile at the end -- I'm assuming that's expected)

Reviewers: @abarciauskas-bgse