QPPA-1588: Updating metricType of measureId 458 (ACR) to be 'cost'
Closes #106

https://jira.cms.gov/browse/QPPA-1588

In QPPA-1588, we'll be adding a new database table to store data for
cost-related measures, like ACR (even though ACR is technically a
quality measure) and future cost measures. This PR changes the
metricType of ACR, which is measure 458, to be costScore and updates our
schemas to expect this as a valid metricType.

Reviewers: @marimiyachi @kencheeto