QPPA-1132: Changing metricType for measure 321 from singlePerformanceRate to cahps
Closes #107

Measure 321 is the aggregate measure for CAHPS and currently has a
metric type of 'singlePerformanceRate'. The aggregate value for CAHPS
score will be a percentage (just like the individual CAHPS measurements)
so we should use the CAHPS metric type for 321 as well. This PR changes
that metricType for measure 321.

Also updating the version number to be 1.0.5 so it can be published and
consumed by downstream services.

Reviewers: @kencheeto @lamroger