QPPA-1606: Ingest 2018 IA measures (#126)
This PR adds a script for ingesting the 2018 IA measures from the source CSV file. This adds the source CSV file, the script that parses that CSV, the staging IA measures, the changes to the build-measures script, and the changes to the final measures. 

Also added some simple tests for IA measures (seems like there weren't any in 2017) as well as tests for the ingestion script and updating the merge-measures tests to use IA measures.

Reviewers: @kalvinwang @kencheeto