QPPA-3147: Remove Nonprop Measures from 2018 Benchmarks (#266)
An incorrect file is replaced with a csv without the non-proportion measures, and a new file for cahps benchmarks is added also. It seems that in a previous PR cahps benchmarks were added directly to the 2018.json file, rather than added as a json file in the staging directory in order to ensure that the benchmarks are added each time the build script is run.