QPPA-3460: Update Cost Measure Descriptions and Measure Specifications (#312)
New URLs have been provided for the measureSpecification field for cost measures. The measureSpecification field has been updated to support the FE's ability to display links for 3 files. The measureSpecification object was updated to include the keys measureInformation, codeList, and changeLog. The import-cost-measures script was updated to properly handle these new keys. Lastly, the 2019 schema was updated to include the keys.