QPPA-3960, QPPSF-5794: Measure spec updates (#327)
#### Motivation for change

Measures 283 and 383 were updated with the correct links as part of QPPA-3960. This PR also includes measure spec link updates for PY 2017 and 2018. PY 2017 source file was updated and the generated file was manually updated as well. A full regeneration of 2017 measures was NOT run.

#### What is being changed

Measure specification links for given measures.

#### Release checklist:

* [x] [Package version updated](https://github.com/CMSgov/qpp-measures-data/blob/master/CONTRIBUTING.md#versioning-publishing-and-creating-new-releases)
* [x] Documentation updated
* [x] Unit tests added/passing
* [x] Verified working locally

##### Associated JIRA tickets:
* https://jira.cms.gov/browse/QPPA-3960
* https://jira.cms.gov/browse/QPPSF-5794